services/api/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
supabase
PyJWT[crypto]
cachetools
